+++ views/map/view.php
@@ -8,17 +8,24 @@
 use humhub\modules\thiscoveryMapping\helpers\Url;
 use humhub\modules\thiscoveryMapping\models\Map;
 use humhub\modules\thiscoveryMapping\widgets\MapWidget;
+use humhub\widgets\bootstrap\Button;
 use yii\helpers\Html;
 
 /** @var Map $map */
 
 $this->title = $map->title;
+$backUrl = Url::toIndex($map->isGlobal() ? null : ($map->content->container ?? null));
 ?>
 <div class="panel panel-default">
-    <div class="panel-heading">
+    <div class="panel-heading tm-view-heading">
         <?= Html::encode($map->title) ?>
-        <?php if ($map->canManage()): ?>
-            <div class="pull-right">
+        <div class="tm-view-heading__actions">
+            <?= Button::light(Yii::t('ThiscoveryMappingModule.base', 'Back to maps'))
+                ->link($backUrl)
+                ->icon('arrow-left')
+                ->sm()
+                ->loader(false) ?>
+            <?php if ($map->canManage()): ?>
                 <a class="btn btn-sm btn-default" href="<?= Html::encode(Url::toEdit($map)) ?>">
                     <?= Yii::t('ThiscoveryMappingModule.base', 'Edit') ?>
                 </a>
@@ -29,8 +36,8 @@ $this->title = $map->title;
                 <?php endif; ?>
                 <a class="btn btn-sm btn-default" href="<?= Html::encode(Url::toExport($map, 'geojson')) ?>">GeoJSON</a>
                 <a class="btn btn-sm btn-default" href="<?= Html::encode(Url::toExport($map, 'csv')) ?>">CSV</a>
-            </div>
-        <?php endif; ?>
+            <?php endif; ?>
+        </div>
     </div>
     <div class="panel-body">
         <?php if ($map->description): ?>